- remove hoary /usr/lib/{rpmpopt,rpmrc} legacy symlinks. - x86_64: remove gcc warning cruft.

+++ popthelp.c
@@ -356,7 +356,7 @@ static void singleOptionHelp(FILE * fp, size_t maxLeftCol,
 /*@=boundswrite@*/
 
     if (help)
-	fprintf(fp,"  %-*s   ", maxLeftCol+displaypad, left);
+	fprintf(fp,"  %-*s   ", (int)(maxLeftCol+displaypad), left);
     else {
 	fprintf(fp,"  %s\n", left); 
 	goto out;
